requirements.txt
pygame-ce==2.5.5
PyTMX==3.32
babel==2.17.0
sphinx
sphinx_rtd_theme